% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushme.R
\name{pushme}
\alias{pushme}
\title{Push a message}
\usage{
pushme(message = NULL, error = FALSE, add_message = NULL)
}
\arguments{
\item{message}{String message to be sent to the pushover app. If none is provided, the function will randomly choose a message from its own list.}

\item{error}{For outputting error messages. If `TRUE`, it will randomly select an error message. If `FALSE` (default), it will randomly select a completion message.}

\item{add_message}{Additional string message that can be added to the main text.}
}
\value{
Random message on your phone/computer.
}
\description{
Push a message to a phone or a computer after is done. It can send random messages when code is complete or there is an error.
}
\examples{
pushme()
pushme(error = TRUE)
pushme("This is my own message")

i = 20
pushme(error = TRUE, add_message = paste("Failure on iteration", i))

}
