% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_pushme.R
\name{set_pushme}
\alias{set_pushme}
\title{Set user and app token}
\usage{
set_pushme(user = "", token = "")
}
\arguments{
\item{user}{User id (string) from pushover app.}

\item{token}{App token (string) from pushover app.}
}
\value{
Random message on your phone/computer.
}
\description{
Wrapper function for `set_pushover_user()` and `set_pushover_app()` to set pushover user and app token. Get a user and token at [https://pushover.net/].
}
\examples{
set_pushme(user = "uQiRzpo4DXghDmr9QzzfQu27cmVRsG", token = "azGDORePK8gMaC0QOYAMyEEuzJnyUi")

}
